+++ src/App.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import HomePage from './pages/HomePage/HomePage';
-import PodcastDetailPage from './pages/PodcastDetailPage/PodcastDetailPage';
-import EpisodeDetailPage from './pages/EpisodeDetailPage/EpisodeDetailPage';
-import Header from './components/Header/Header';
-import { LoadingProvider } from './context/LoadingContext';
-import { NavigationProvider } from './context/NavigationContext';
-import { PodcastServiceProvider } from './context/PodcastServiceContext'; // Importa el proveedor
+import HomePage from './features/podcasts/pages/HomePage/HomePage';
+import PodcastDetailPage from './features/podcasts/pages/PodcastDetailPage/PodcastDetailPage';
+import EpisodeDetailPage from './features/podcasts/pages/EpisodeDetailPage/EpisodeDetailPage';
+import Header from './shared/components/Header';
+import { LoadingProvider } from './shared/context/LoadingContext';
+import { NavigationProvider } from './shared/context/NavigationContext';
+import { PodcastServiceProvider } from './shared/context/PodcastServiceContext'; // Importa el proveedor
 
 
 const App: React.FC = () => {
@@ -31,11 +31,10 @@ const App: React.FC = () => {
               {/* Ruta para la página principal */}
               <Route path="/" element={<HomePage />} />
 
-              {/* Ruta para el detalle de un podcast */}
-              <Route path="/podcast/:podcastId" element={<PodcastDetailPage />} />
-
               {/* Ruta para el detalle de un episodio */}
-              <Route path="/podcast/:podcastId/episode/:episodeId" element={<EpisodeDetailPage />} />
+              <Route path="/podcast/:podcastId" element={<PodcastDetailPage />}>
+                <Route path="episode/:episodeId" element={<EpisodeDetailPage />} />
+              </Route>
 
             </Routes>
           </PodcastServiceProvider>
