+++ src/features/podcasts/components/PodcastCard.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import './../../styles/card.scss';
+import './../styles/card.scss';
 
 
 interface CardProps {
