+++ src/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-import './styles/global.scss';
+import './../src/shared/styles/global.scss';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
