+++ src/shared/components/Header.tsx
@@ -1,14 +1,15 @@
 import React from 'react';
-import { Link } from 'react-router-dom';
-import LoadingIndicator from '../LoadingIndicator/LoadingIndicator';
-import { useLoading } from '../../context/LoadingContext';
-import useHomeNavigation from '../../hook/useHomeNavigation';
-import './../../styles/header.scss';
+import { Link, useNavigate } from 'react-router-dom';
+import LoadingIndicator from './LoadingIndicator';
+import { useLoading } from './../context/LoadingContext';
+import useHomeNavigation from './../hooks/useHomeNavigation';
+import './../styles/header.scss';
 
 
 function Header() {
     const { handleHomeClick } = useHomeNavigation();
     const { loading } = useLoading(); // Acceder al estado de carga
+    const navigate = useNavigate(); // Hook for navigation
 
 
     return (
@@ -19,7 +20,8 @@ function Header() {
                 onClick={(e) => {
                     e.preventDefault();
                     handleHomeClick();
-                } }
+                    navigate("/");
+                }}
             >
                 Podcaster
             </Link>
