+++ src/shared/styles/header.scss
