+++ src/shared/styles/loadingIndicator.scss
